+++ phillydb/log_config.py
@@ -1,6 +1,0 @@
-import logging
-import sys
-
-
-logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
-logger = logging.getLogger(__name__)