+++ tests/test_labelled_fields.py
@@ -1,0 +1,10 @@
+import pandas as pd
+import pytest
+
+from phillydb.labelled_fields import label_city_owned_properties
+
+def test_label_city_owned_properties():
+    with pytest.raises(ValueError):
+        label_city_owned_properties(pd.DataFrame())
+
+